+++ HISTORY.md
@@ -15,6 +15,8 @@ Add support for searching observations by observation fields, using a new `obser
 * `get_observation_species_counts()`
 
 ### Models
+* Add `Observation.ident_taxon_ids` dynamic property to get all identification taxon IDs (with ancestors)
+* Add `Observation.cumulative_ids` dynamic property to calculate agreements/total community identifications
 * Add `Sound` model for `Observation.sounds`
 * Fix initialization of `ListedTaxon.place`